Remove version overriding stuff unless necessary
By doing this we avoid circular requires (`rubygems` requires `bundler` via
`USE_GEMDEPS`, `bundler` requires `rubygems` when loading its own version), and
also redefinition warnings when the `bundler.gemspec` is evaluated with
`rubygems` already loaded (for example, when running `ruby setup.rb`
from `rubygems` repo).

When working with rubygems versions that didn't have the bundler version
finder, it seems to be necessary, but for newer rubygems versions it's
not.